Fix category/label type hints from int to float
classify_state()/AssociationNode.value() always produce float 0.0/1.0, but
random_alternating_training_data's return type, train_linear_classifier_network
and plot_training_data's training_data parameter, and
LinearClassifierNetwork.learn's category parameter were all annotated as
int. Harmless at runtime (Python doesn't enforce hints, and int compares
fine against float), but inaccurate for a reader or type checker.

+++ perceptron/graphics/chart.py
@@ -28,7 +28,7 @@ def plot_linear_classifier_network(
         axes.add_line(line_graph)
 
 
-def plot_training_data(axes: Axes, training_data: list[tuple[tuple[float, float], int]]):
+def plot_training_data(axes: Axes, training_data: list[tuple[tuple[float, float], float]]):
 
     markers = [".", "x"]
     colors = ["blue", "yellow"]
